luci: fix app update some bug

+++ luci-app-passwall/luasrc/model/cbi/passwall/api/brook.lua
@@ -8,6 +8,13 @@ local api = require "luci.model.cbi.passwall.api.api"
 local brook_api = "https://api.github.com/repos/txthinking/brook/releases/latest"
 
 function to_check(arch)
+    local app_path = api.get_brook_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Brook")
+        }
+    end
     if not arch or arch == "" then arch = api.auto_get_arch() end
 
     local file_tree, sub_version = api.get_file_info(arch)
@@ -15,8 +22,7 @@ function to_check(arch)
     if file_tree == "" then
         return {
             code = 1,
-            error = i18n.translate(
-                "Can't determine ARCH, or ARCH not supported.")
+            error = i18n.translate("Can't determine ARCH, or ARCH not supported.")
         }
     end
 
@@ -66,6 +72,13 @@ function to_check(arch)
 end
 
 function to_download(url)
+    local app_path = api.get_brook_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Brook")
+        }
+    end
     if not url or url == "" then
         return {code = 1, error = i18n.translate("Download url is required.")}
     end
@@ -88,13 +101,20 @@ function to_download(url)
 end
 
 function to_move(file)
+    local app_path = api.get_brook_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Brook")
+        }
+    end
     if not file or file == "" or not fs.access(file) then
         sys.call("/bin/rm -rf /tmp/brook_download.*")
         return {code = 1, error = i18n.translate("Client file is required.")}
     end
 
-    local version = api.get_brook_version(file)
-    if version == "" then
+    local new_version = api.get_brook_version(file)
+    if new_version == "" then
         sys.call("/bin/rm -rf /tmp/brook_download.*")
         return {
             code = 1,
@@ -102,30 +122,29 @@ function to_move(file)
         }
     end
 
-    local client_file = api.get_brook_path()
-    local client_file_bak
+    local app_path_bak
 
-    if fs.access(client_file) then
-        client_file_bak = client_file .. ".bak"
-        api.exec("/bin/mv", {"-f", client_file, client_file_bak})
+    if fs.access(app_path) then
+        app_path_bak = app_path .. ".bak"
+        api.exec("/bin/mv", {"-f", app_path, app_path_bak})
     end
 
-    local result = api.exec("/bin/mv", {"-f", file, client_file}, nil, api.command_timeout) == 0
+    local result = api.exec("/bin/mv", {"-f", file, app_path}, nil, api.command_timeout) == 0
 
-    if not result or not fs.access(client_file) then
+    if not result or not fs.access(app_path) then
         sys.call("/bin/rm -rf /tmp/brook_download.*")
-        if client_file_bak then
-            api.exec("/bin/mv", {"-f", client_file_bak, client_file})
+        if app_path_bak then
+            api.exec("/bin/mv", {"-f", app_path_bak, app_path})
         end
         return {
             code = 1,
-            error = i18n.translatef("Can't move new file to path: %s", client_file)
+            error = i18n.translatef("Can't move new file to path: %s", app_path)
         }
     end
 
-    api.exec("/bin/chmod", {"755", client_file})
+    api.exec("/bin/chmod", {"755", app_path})
 
-    if client_file_bak then api.exec("/bin/rm", {"-f", client_file_bak}) end
+    if app_path_bak then api.exec("/bin/rm", {"-f", app_path_bak}) end
 
     sys.call("/bin/rm -rf /tmp/brook_download.*")
