+++ luci-app-passwall/luasrc/model/cbi/passwall/api/kcptun.lua
@@ -7,30 +7,14 @@ local api = require "luci.model.cbi.passwall.api.api"
 
 local kcptun_api = "https://api.github.com/repos/xtaci/kcptun/releases/latest"
 
-function get_kcptun_file_path()
-    return api.uci_get_type("global_app", "kcptun_client_file")
-end
-
-function get_kcptun_version(file)
-    if file == nil then file = get_kcptun_file_path() end
-
-    if file and file ~= "" then
-        if not fs.access(file, "rwx", "rx", "rx") then
-            fs.chmod(file, 755)
-        end
-
-        local info = util.trim(sys.exec("%s -v 2>/dev/null" % file))
-
-        if info ~= "" then
-            local tb = util.split(info, "%s+", nil, true)
-            return tb[1] == "kcptun" and tb[3] or ""
-        end
-    end
-
-    return ""
-end
-
 function to_check(arch)
+    local app_path = api.get_kcptun_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Kcptun")
+        }
+    end
     if not arch or arch == "" then arch = api.auto_get_arch() end
 
     local file_tree, sub_version = api.get_file_info(arch)
@@ -38,8 +22,7 @@ function to_check(arch)
     if file_tree == "" then
         return {
             code = 1,
-            error = i18n.translate(
-                "Can't determine ARCH, or ARCH not supported.")
+            error = i18n.translate("Can't determine ARCH, or ARCH not supported.")
         }
     end
 
@@ -52,12 +35,9 @@ function to_check(arch)
         }
     end
 
+    local now_version = api.get_kcptun_version()
     local remote_version = json.tag_name:match("[^v]+")
-
-    local client_file = get_kcptun_file_path()
-
-    local needs_update = api.compare_versions(get_kcptun_version(client_file),
-                                              "<", remote_version)
+    local needs_update = api.compare_versions(now_version, "<", remote_version)
     local html_url, download_url
 
     if needs_update then
@@ -73,24 +53,30 @@ function to_check(arch)
     if needs_update and not download_url then
         return {
             code = 1,
-            now_version = get_kcptun_version(client_file),
+            now_version = now_version,
             version = remote_version,
             html_url = html_url,
-            error = i18n.translate(
-                "New version found, but failed to get new version download url.")
+            error = i18n.translate("New version found, but failed to get new version download url.")
         }
     end
 
     return {
         code = 0,
         update = needs_update,
-        now_version = get_kcptun_version(client_file),
+        now_version = now_version,
         version = remote_version,
         url = {html = html_url, download = download_url}
     }
 end
 
 function to_download(url)
+    local app_path = api.get_kcptun_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Kcptun")
+        }
+    end
     if not url or url == "" then
         return {code = 1, error = i18n.translate("Download url is required.")}
     end
@@ -113,6 +99,13 @@ function to_download(url)
 end
 
 function to_extract(file, subfix)
+    local app_path = api.get_kcptun_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Kcptun")
+        }
+    end
     if not file or file == "" or not fs.access(file) then
         return {code = 1, error = i18n.translate("File path required.")}
     end
@@ -157,47 +150,50 @@ function to_extract(file, subfix)
 end
 
 function to_move(file)
+    local app_path = api.get_kcptun_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Kcptun")
+        }
+    end
     if not file or file == "" or not fs.access(file) then
         sys.call("/bin/rm -rf /tmp/kcptun_extract.*")
         return {code = 1, error = i18n.translate("Client file is required.")}
     end
 
-    local version = get_kcptun_version(file)
-    if version == "" then
+    local new_version = api.get_kcptun_version(file)
+    if new_version == "" then
         sys.call("/bin/rm -rf /tmp/kcptun_extract.*")
         return {
             code = 1,
-            error = i18n.translate(
-                "The client file is not suitable for current device.")
+            error = i18n.translate("The client file is not suitable for current device.")
         }
     end
 
-    local client_file = get_kcptun_file_path()
-    local client_file_bak
+    local app_path_bak
 
-    if fs.access(client_file) then
-        client_file_bak = client_file .. ".bak"
-        api.exec("/bin/mv", {"-f", client_file, client_file_bak})
+    if fs.access(app_path) then
+        app_path_bak = app_path .. ".bak"
+        api.exec("/bin/mv", {"-f", app_path, app_path_bak})
     end
 
-    local result = api.exec("/bin/mv", {"-f", file, client_file}, nil,
-                            api.command_timeout) == 0
+    local result = api.exec("/bin/mv", {"-f", file, app_path}, nil, api.command_timeout) == 0
 
-    if not result or not fs.access(client_file) then
+    if not result or not fs.access(app_path) then
         sys.call("/bin/rm -rf /tmp/kcptun_extract.*")
-        if client_file_bak then
-            api.exec("/bin/mv", {"-f", client_file_bak, client_file})
+        if app_path_bak then
+            api.exec("/bin/mv", {"-f", app_path_bak, app_path})
         end
         return {
             code = 1,
-            error = i18n.translatef("Can't move new file to path: %s",
-                                    client_file)
+            error = i18n.translatef("Can't move new file to path: %s", app_path)
         }
     end
 
-    api.exec("/bin/chmod", {"755", client_file})
+    api.exec("/bin/chmod", {"755", app_path})
 
-    if client_file_bak then api.exec("/bin/rm", {"-f", client_file_bak}) end
+    if app_path_bak then api.exec("/bin/rm", {"-f", app_path_bak}) end
 
     sys.call("/bin/rm -rf /tmp/kcptun_extract.*")
