+++ luci-app-passwall/luasrc/model/cbi/passwall/api/trojan_go.lua
@@ -3,12 +3,18 @@ local fs = require "nixio.fs"
 local sys = require "luci.sys"
 local util = require "luci.util"
 local i18n = require "luci.i18n"
-local ipkg = require("luci.model.ipkg")
 local api = require "luci.model.cbi.passwall.api.api"
 
 local trojan_go_api = api.uci_get_type("global_app", "trojan_go_latest", "https://api.github.com/repos/trojan-gfw/trojan-go/releases/latest")
 
 function to_check(arch)
+    local app_path = api.get_trojan_go_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Trojan-GO")
+        }
+    end
     if not arch or arch == "" then arch = api.auto_get_arch() end
 
     local file_tree, sub_version = api.get_file_info(arch)
@@ -16,8 +22,7 @@ function to_check(arch)
     if file_tree == "" then
         return {
             code = 1,
-            error = i18n.translate(
-                "Can't determine ARCH, or ARCH not supported.")
+            error = i18n.translate("Can't determine ARCH, or ARCH not supported.")
         }
     end
 
@@ -59,8 +64,7 @@ function to_check(arch)
             now_version = now_version,
             version = remote_version,
             html_url = html_url,
-            error = i18n.translate(
-                "New version found, but failed to get new version download url.") .. " [linux-" .. file_tree .. ".zip]"
+            error = i18n.translate("New version found, but failed to get new version download url.") .. " [linux-" .. file_tree .. ".zip]"
         }
     end
 
@@ -74,6 +78,13 @@ function to_check(arch)
 end
 
 function to_download(url)
+    local app_path = api.get_trojan_go_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Trojan-GO")
+        }
+    end
     if not url or url == "" then
         return {code = 1, error = i18n.translate("Download url is required.")}
     end
@@ -96,10 +107,19 @@ function to_download(url)
 end
 
 function to_extract(file, subfix)
-    local isinstall_unzip = ipkg.installed("unzip")
-    if isinstall_unzip == nil then
-        ipkg.update()
-        ipkg.install("unzip")
+    local app_path = api.get_trojan_go_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Trojan-GO")
+        }
+    end
+    if sys.exec("echo -n $(opkg list-installed | grep -c unzip)") ~= "1" then
+        api.exec("/bin/rm", {"-f", file})
+        return {
+            code = 1,
+            error = i18n.translate("Not installed unzip, Can't unzip!")
+        }
     end
 
     if not file or file == "" or not fs.access(file) then
@@ -121,29 +141,35 @@ function to_extract(file, subfix)
 end
 
 function to_move(file)
+    local app_path = api.get_trojan_go_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Trojan-GO")
+        }
+    end
     if not file or file == "" then
         sys.call("/bin/rm -rf /tmp/trojan-go_extract.*")
         return {code = 1, error = i18n.translate("Client file is required.")}
     end
 
-    local client_file = api.get_trojan_go_path()
-    local client_file_bak
+    local app_path_bak
 
-    if fs.access(client_file) then
-        client_file_bak = client_file .. ".bak"
-        api.exec("/bin/mv", {"-f", client_file, client_file_bak})
+    if fs.access(app_path) then
+        app_path_bak = app_path .. ".bak"
+        api.exec("/bin/mv", {"-f", app_path, app_path_bak})
     end
 
-    local result = api.exec("/bin/mv", { "-f", file .. "/trojan-go", client_file }, nil, api.command_timeout) == 0
+    local result = api.exec("/bin/mv", { "-f", file .. "/trojan-go", app_path }, nil, api.command_timeout) == 0
     sys.call("/bin/rm -rf /tmp/trojan-go_extract.*")
-    if not result or not fs.access(client_file) then
+    if not result or not fs.access(app_path) then
         return {
             code = 1,
-            error = i18n.translatef("Can't move new file to path: %s", client_file)
+            error = i18n.translatef("Can't move new file to path: %s", app_path)
         }
     end
 
-    api.exec("/bin/chmod", {"-R", "755", client_file})
+    api.exec("/bin/chmod", {"-R", "755", app_path})
 
     return {code = 0}
 end
