+++ luci-app-passwall/luasrc/model/cbi/passwall/api/xray.lua
@@ -3,26 +3,19 @@ local fs = require "nixio.fs"
 local sys = require "luci.sys"
 local util = require "luci.util"
 local i18n = require "luci.i18n"
-local ipkg = require("luci.model.ipkg")
 local api = require "luci.model.cbi.passwall.api.api"
 
 local xray_api = "https://api.github.com/repos/XTLS/Xray-core/releases/latest"
 local is_armv7 = false
 
-function get_xray_file_path()
-    return api.uci_get_type("global_app", "xray_file")
-end
-
-function get_xray_version()
-    if get_xray_file_path() and get_xray_file_path() ~= "" then
-        if fs.access(get_xray_file_path() .. "/xray") then
-            return sys.exec("echo -n $(" .. get_xray_file_path() .. "/xray -version | awk '{print $2}' | sed -n 1P" .. ")")
-        end
-    end
-    return ""
-end
-
 function to_check(arch)
+    local app_path = api.get_xray_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Xray")
+        }
+    end
     if not arch or arch == "" then arch = api.auto_get_arch() end
 
     local file_tree, sub_version = api.get_file_info(arch)
@@ -31,8 +24,7 @@ function to_check(arch)
     if file_tree == "" then
         return {
             code = 1,
-            error = i18n.translate(
-                "Can't determine ARCH, or ARCH not supported.")
+            error = i18n.translate("Can't determine ARCH, or ARCH not supported.")
         }
     end
 
@@ -48,8 +40,9 @@ function to_check(arch)
         }
     end
 
+    local now_version = api.get_xray_version()
     local remote_version = json.tag_name:match("[^v]+")
-    local needs_update = api.compare_versions(get_xray_version(), "<", remote_version)
+    local needs_update = api.compare_versions(now_version, "<", remote_version)
     local html_url, download_url
 
     if needs_update then
@@ -65,24 +58,30 @@ function to_check(arch)
     if needs_update and not download_url then
         return {
             code = 1,
-            now_version = get_xray_version(),
+            now_version = now_version,
             version = remote_version,
             html_url = html_url,
-            error = i18n.translate(
-                "New version found, but failed to get new version download url.")
+            error = i18n.translate("New version found, but failed to get new version download url.")
         }
     end
 
     return {
         code = 0,
         update = needs_update,
-        now_version = get_xray_version(),
+        now_version = now_version,
         version = remote_version,
         url = {html = html_url, download = download_url}
     }
 end
 
 function to_download(url)
+    local app_path = api.get_xray_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Xray")
+        }
+    end
     if not url or url == "" then
         return {code = 1, error = i18n.translate("Download url is required.")}
     end
@@ -105,10 +104,19 @@ function to_download(url)
 end
 
 function to_extract(file, subfix)
-    local isinstall_unzip = ipkg.installed("unzip")
-    if isinstall_unzip == nil then
-        ipkg.update()
-        ipkg.install("unzip")
+    local app_path = api.get_xray_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Xray")
+        }
+    end
+    if sys.exec("echo -n $(opkg list-installed | grep -c unzip)") ~= "1" then
+        api.exec("/bin/rm", {"-f", file})
+        return {
+            code = 1,
+            error = i18n.translate("Not installed unzip, Can't unzip!")
+        }
     end
 
     if not file or file == "" or not fs.access(file) then
@@ -130,37 +138,38 @@ function to_extract(file, subfix)
 end
 
 function to_move(file)
+    local app_path = api.get_xray_path() or ""
+    if app_path == "" then
+        return {
+            code = 1,
+            error = i18n.translatef("You did not fill in the %s path. Please save and apply then update manually.", "Xray")
+        }
+    end
     if not file or file == "" then
         sys.call("/bin/rm -rf /tmp/xray_extract.*")
         return {code = 1, error = i18n.translate("Client file is required.")}
     end
 
-    local client_file = get_xray_file_path()
-
-    sys.call("mkdir -p " .. client_file)
+    sys.call("mkdir -p " .. app_path)
 
     if not arch or arch == "" then arch = api.auto_get_arch() end
     local file_tree, sub_version = api.get_file_info(arch)
     if sub_version == "7" then is_armv7 = true end
     local result = nil
     if is_armv7 and is_armv7 == true then
-        result = api.exec("/bin/mv", {
-            "-f", file .. "/xray_armv7", file .. "/v2ctl_armv7", client_file
-        }, nil, api.command_timeout) == 0
+        result = api.exec("/bin/mv", { "-f", file .. "/xray_armv7", app_path }, nil, api.command_timeout) == 0
     else
-        result = api.exec("/bin/mv", {
-            "-f", file .. "/xray", file .. "/v2ctl", client_file
-        }, nil, api.command_timeout) == 0
+        result = api.exec("/bin/mv", { "-f", file .. "/xray", app_path }, nil, api.command_timeout) == 0
     end
     sys.call("/bin/rm -rf /tmp/xray_extract.*")
-    if not result or not fs.access(client_file) then
+    if not result or not fs.access(app_path) then
         return {
             code = 1,
-            error = i18n.translatef("Can't move new file to path: %s", client_file)
+            error = i18n.translatef("Can't move new file to path: %s", app_path)
         }
     end
 
-    api.exec("/bin/chmod", {"-R", "755", client_file})
+    api.exec("/bin/chmod", {"-R", "755", app_path})
 
     return {code = 0}
 end
